removed use of String in com outputs used F() macro program size now down to 25774 bytes of flash verified still will not run on uno

+++ AnaloguePin.cpp
@@ -14,7 +14,7 @@ AnaloguePin::AnaloguePin() {
     factor = 0;
 }
 
-void AnaloguePin::set(int inPin, float conversionFactor) {
+void AnaloguePin::set(byte inPin, float conversionFactor) {
     pinNo = inPin;
     factor = conversionFactor;
 }
